js/xpconnect/tests/mochitest/test_bug871887.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=871887
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 871887</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 871887 **/
  SimpleTest.waitForExplicitFinish();

  // NB: onstart ends up getting invoked twice, for mysterious and potentially-
  // IE6-related reasons.
  function checkpoint(invocant) {
    ok(true, "onstart called");
    is(invocant, $('llama'), "this-binding is correct");
    $('llama').loop = 1;
    $('llama').scrollDelay = 1;
    $('llama').scrollAmount = 500;
  }

  function done(invocant) {
    is(invocant, $('llama'), "this-binding is correct");
    ok(true, "onfinish called");
    SimpleTest.finish();
  }

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=871887">Mozilla Bug 871887</a>
<p id="display"></p>
<div id="content">
<marquee id="llama" onstart="checkpoint(this);" onfinish="done(this);">Watch the Llama</marquee>
</div>
<pre id="test">
</pre>
</body>
</html>
